+++ src/org/jitsi/impl/neomedia/RTPConnectorOutputStream.java
@@ -10,7 +10,6 @@ import java.io.*;
 import java.net.*;
 import java.util.*;
 import java.util.concurrent.*;
-import java.util.concurrent.locks.*;
 
 import javax.media.rtp.*;
 
@@ -35,16 +34,32 @@ public abstract class RTPConnectorOutputStream
         = Logger.getLogger(RTPConnectorOutputStream.class);
 
     /**
-     * The maximum number of packets to be sent to be kept in the queue of
-     * <tt>MaxPacketsPerMillisPolicy</tt>. When the maximum is reached, the next
-     * attempt to write a new packet in the queue will block until at least one
-     * packet from the queue is sent. Defined in order to prevent
-     * <tt>OutOfMemoryError</tt>s which, technically, may arise if the capacity
-     * of the queue is unlimited.
+     * Determines whether a <tt>RawPacket</tt> which has a specific number in
+     * the total number of sent <tt>RawPacket</tt>s is to be logged by
+     * {@link PacketLoggingService}.
+     *
+     * @param numOfPacket the number of the <tt>RawPacket</tt> in the total
+     * number of sent <tt>RawPacket</tt>s
+     * @return <tt>true</tt> if the <tt>RawPacket</tt> with the specified
+     * <tt>numOfPacket</tt> is to be logged by <tt>PacketLoggingService</tt>;
+     * otherwise, <tt>false</tt>
      */
-    public static final int
-        MAX_PACKETS_PER_MILLIS_POLICY_PACKET_QUEUE_CAPACITY
-            = 256;
+    static boolean logPacket(long numOfPacket)
+    {
+        return
+            (numOfPacket == 1)
+                || (numOfPacket == 300)
+                || (numOfPacket == 500)
+                || (numOfPacket == 1000)
+                || ((numOfPacket % 5000) == 0);
+    }
+
+    /**
+     * Whether this <tt>RTPConnectorOutputStream</tt> is enabled or disabled.
+     * While the stream is disabled, it suppresses actually sending any packets
+     * via {@link #send(RawPacket)}.
+     */
+    private boolean enabled = true;
 
     /**
      * The functionality which allows this <tt>OutputDataStream</tt> to control
@@ -54,17 +69,15 @@ public abstract class RTPConnectorOutputStream
     private MaxPacketsPerMillisPolicy maxPacketsPerMillisPolicy;
 
     /**
-     * Stream targets' IP addresses and ports.
+     * Number of bytes sent through this stream.
      */
-    protected final List<InetSocketAddress> targets
-        = new LinkedList<InetSocketAddress>();
+    private long numberOfBytesSent = 0;
 
     /**
-     * The pool of <tt>RawPacket</tt> instances which reduces the number of
-     * allocations performed by {@link #createRawPacket(byte[], int, int)}.
+     * Used for debugging. As we don't log every packet
+     * we must count them and decide which to log.
      */
-    private final LinkedBlockingQueue<RawPacket> rawPacketPool
-        = new LinkedBlockingQueue<RawPacket>();
+    private long numberOfPackets = 0;
 
     /**
      * The pool of <tt>RawPacket[]</tt> instances which reduces the number of
@@ -75,22 +88,17 @@ public abstract class RTPConnectorOutputStream
         = new LinkedBlockingQueue<RawPacket[]>();
 
     /**
-     * Used for debugging. As we don't log every packet
-     * we must count them and decide which to log.
+     * The pool of <tt>RawPacket</tt> instances which reduces the number of
+     * allocations performed by {@link #createRawPacket(byte[], int, int)}.
      */
-    private long numberOfPackets = 0;
+    private final LinkedBlockingQueue<RawPacket> rawPacketPool
+        = new LinkedBlockingQueue<RawPacket>();
 
     /**
-     * Whether this <tt>RTPConnectorOutputStream</tt> is enabled or disabled.
-     * While the stream is disabled, it suppresses actually sending any packets
-     * via {@link #send(RawPacket)}.
+     * Stream targets' IP addresses and ports.
      */
-    private boolean enabled = true;
-
-    /**
-     * Number of bytes sent through this stream.
-     */
-    private long numberOfBytesSent = 0;
+    protected final List<InetSocketAddress> targets
+        = new LinkedList<InetSocketAddress>();
 
     /**
      * Initializes a new <tt>RTPConnectorOutputStream</tt> which is to send
@@ -113,15 +121,6 @@ public abstract class RTPConnectorOutputStream
 
         if (!targets.contains(target))
             targets.add(target);
-    }
-
-    /**
-     * Returns the number of bytes sent trough this stream
-     * @return the number of bytes sent
-     */
-    public long getNumberOfBytesSent()
-    {
-        return numberOfBytesSent;
     }
 
     /**
@@ -193,6 +192,35 @@ public abstract class RTPConnectorOutputStream
     }
 
     /**
+     * Logs a specific <tt>RawPacket</tt> associated with a specific remote
+     * address.
+     *
+     * @param packet packet to log
+     * @param target the remote address associated with the <tt>packet</tt>
+     */
+    protected abstract void doLogPacket(
+        RawPacket packet,
+        InetSocketAddress target);
+
+    /**
+     * Returns the number of bytes sent trough this stream
+     * @return the number of bytes sent
+     */
+    public long getNumberOfBytesSent()
+    {
+        return numberOfBytesSent;
+    }
+
+    /**
+     * Returns whether or not this <tt>RTPConnectorOutputStream</tt> has a valid
+     * socket.
+     *
+     * @return <tt>true</tt> if this <tt>RTPConnectorOutputStream</tt> has a
+     * valid socket; <tt>false</tt>, otherwise
+     */
+    protected abstract boolean isSocketValid();
+
+    /**
      * Remove a target from stream targets list
      *
      * @param remoteAddr target ip address
@@ -224,64 +252,6 @@ public abstract class RTPConnectorOutputStream
     {
         targets.clear();
     }
-
-    /**
-     * Determines whether a <tt>RawPacket</tt> which has a specific number in
-     * the total number of sent <tt>RawPacket</tt>s is to be logged by
-     * {@link PacketLoggingService}.
-     *
-     * @param numOfPacket the number of the <tt>RawPacket</tt> in the total
-     * number of sent <tt>RawPacket</tt>s
-     * @return <tt>true</tt> if the <tt>RawPacket</tt> with the specified
-     * <tt>numOfPacket</tt> is to be logged by <tt>PacketLoggingService</tt>;
-     * otherwise, <tt>false</tt>
-     */
-    static boolean logPacket(long numOfPacket)
-    {
-        return
-            (numOfPacket == 1)
-                || (numOfPacket == 300)
-                || (numOfPacket == 500)
-                || (numOfPacket == 1000)
-                || ((numOfPacket % 5000) == 0);
-    }
-
-    /**
-     * Sends a specific <tt>RawPacket</tt> through this
-     * <tt>OutputDataStream</tt> to a specific <tt>InetSocketAddress</tt>.
-     *
-     * @param packet the <tt>RawPacket</tt> to send through this
-     * <tt>OutputDataStream</tt> to the specified <tt>target</tt>
-     * @param target the <tt>InetSocketAddress</tt> to which the specified
-     * <tt>packet</tt> is to be sent through this <tt>OutputDataStream</tt>
-     * @throws IOException if anything goes wrong while sending the specified
-     * <tt>packet</tt> through this <tt>OutputDataStream</tt> to the specified
-     * <tt>target</tt>
-     */
-    protected abstract void sendToTarget(
-            RawPacket packet,
-            InetSocketAddress target)
-        throws IOException;
-
-    /**
-     * Logs a specific <tt>RawPacket</tt> associated with a specific remote
-     * address.
-     *
-     * @param packet packet to log
-     * @param target the remote address associated with the <tt>packet</tt>
-     */
-    protected abstract void doLogPacket(
-        RawPacket packet,
-        InetSocketAddress target);
-
-    /**
-     * Returns whether or not this <tt>RTPConnectorOutputStream</tt> has a valid
-     * socket.
-     *
-     * @return <tt>true</tt> if this <tt>RTPConnectorOutputStream</tt> has a
-     * valid socket; <tt>false</tt>, otherwise
-     */
-    protected abstract boolean isSocketValid();
 
     /**
      * Sends a specific RTP packet through the <tt>DatagramSocket</tt> of this
@@ -332,6 +302,38 @@ public abstract class RTPConnectorOutputStream
     }
 
     /**
+     * Sends a specific <tt>RawPacket</tt> through this
+     * <tt>OutputDataStream</tt> to a specific <tt>InetSocketAddress</tt>.
+     *
+     * @param packet the <tt>RawPacket</tt> to send through this
+     * <tt>OutputDataStream</tt> to the specified <tt>target</tt>
+     * @param target the <tt>InetSocketAddress</tt> to which the specified
+     * <tt>packet</tt> is to be sent through this <tt>OutputDataStream</tt>
+     * @throws IOException if anything goes wrong while sending the specified
+     * <tt>packet</tt> through this <tt>OutputDataStream</tt> to the specified
+     * <tt>target</tt>
+     */
+    protected abstract void sendToTarget(
+            RawPacket packet,
+            InetSocketAddress target)
+        throws IOException;
+
+    /**
+     * Enables or disables this <tt>RTPConnectorOutputStream</tt>.
+     * While the stream is disabled, it suppresses actually sending any packets
+     * via {@link #send(RawPacket)}.
+     *
+     * @param enabled <tt>true</tt> to enable, <tt>false</tt> to disable.
+     */
+    public void setEnabled(boolean enabled)
+    {
+        if (logger.isDebugEnabled())
+            logger.debug("setEnabled: " + enabled);
+
+        this.enabled = enabled;
+    }
+
+    /**
      * Sets the maximum number of RTP packets to be sent by this
      * <tt>OutputDataStream</tt> through its <tt>DatagramSocket</tt> per
      * a specific number of milliseconds.
@@ -353,7 +355,17 @@ public abstract class RTPConnectorOutputStream
                     throw new IllegalArgumentException("perMillis");
 
                 maxPacketsPerMillisPolicy
-                    = new MaxPacketsPerMillisPolicy(maxPackets, perMillis);
+                    = new MaxPacketsPerMillisPolicy(maxPackets, perMillis)
+                    {
+                        /**
+                         * {@inheritDoc}
+                         */
+                        @Override
+                        protected void send(RawPacket packet)
+                        {
+                            RTPConnectorOutputStream.this.send(packet);
+                        }
+                    };
             }
         }
         else
@@ -361,6 +373,15 @@ public abstract class RTPConnectorOutputStream
             maxPacketsPerMillisPolicy
                 .setMaxPacketsPerMillis(maxPackets, perMillis);
         }
+    }
+
+    /**
+     * Changes current thread priority.
+     * @param priority the new priority.
+     */
+    public void setPriority(int priority)
+    {
+        // currently no priority is set
     }
 
     /**
@@ -377,300 +398,58 @@ public abstract class RTPConnectorOutputStream
      */
     public int write(byte[] buffer, int offset, int length)
     {
-        /*
-         * While calling write without targets can be carried out without a
-         * problem, such a situation may be a symptom of a problem. For example,
-         * it was discovered during testing that RTCP was seemingly-endlessly
-         * sent after hanging up a call.
-         */
+        // While calling write without targets can be carried out without a
+        // problem, such a situation may be a symptom of a problem. For example,
+        // it was discovered during testing that RTCP was seemingly-endlessly
+        // sent after hanging up a call.
         if (logger.isDebugEnabled() && targets.isEmpty())
             logger.debug("Write called without targets!", new Throwable());
 
-        // no need to handle the buffer at all, if we are disabled, but simulate
-        // a successful operation.
+        // No need to handle the buffer at all if we are disabled apart from
+        // simulating a successful operation.
         if (!enabled)
             return length;
 
         // get the array of RawPackets we need to send
         RawPacket[] pkts = createRawPacket(buffer, offset, length);
         boolean fail = false;
+
         for(int i = 0; i < pkts.length; i++)
         {
             RawPacket pkt = pkts[i];
-            pkts[i] = null; //clear the array before returning to the pool
-            /*
-             * If we got extended, the delivery of the packet may have been
-             * canceled.
-             */
-            if (pkt != null && !fail)
+
+            pkts[i] = null; // Clear the array before returning to the pool.
+
+            // If we got extended, the delivery of the packet may have been
+            // canceled.
+            if (pkt != null)
             {
-                if (maxPacketsPerMillisPolicy == null)
+                if (fail)
                 {
-                    if (!send(pkt))
-                    {
-                        //skip sending the rest, but return them to the pool
-                        fail = true;
-                        continue;
-                    }
+                    rawPacketPool.offer(pkt);
                 }
                 else
                 {
-                    maxPacketsPerMillisPolicy.write(pkt);
+                    if (maxPacketsPerMillisPolicy == null)
+                    {
+                        if (!send(pkt))
+                        {
+                            // Skip sending the remaining RawPackets but return
+                            // them to the pool and clear pkts. The current pkt
+                            // was returned to the pool.
+                            fail = true;
+                        }
+                    }
+                    else
+                    {
+                        maxPacketsPerMillisPolicy.write(pkt);
+                    }
                 }
             }
-
-            if (pkt != null && fail) //if !fail, send() will have returned it
-                rawPacketPool.offer(pkt);
         }
 
         rawPacketArrayPool.offer(pkts);
 
         return fail ? -1 : length;
-    }
-
-    /**
-     * Changes current thread priority.
-     * @param priority the new priority.
-     */
-    public void setPriority(int priority)
-    {
-        // currently no priority is set
-//        if ((maxPacketsPerMillisPolicy != null)
-//                && (maxPacketsPerMillisPolicy.sendThread != null))
-//            maxPacketsPerMillisPolicy.sendThread.setPriority(priority);
-    }
-
-    /**
-     * Implements the functionality which allows this <tt>OutputDataStream</tt>
-     * to control how many RTP packets it sends through its
-     * <tt>DatagramSocket</tt> per a specific number of milliseconds.
-     */
-    private class MaxPacketsPerMillisPolicy
-    {
-
-        /**
-         * The maximum number of RTP packets to be sent by this
-         * <tt>OutputDataStream</tt> through its <tt>DatagramSocket</tt> per
-         * {@link #perNanos} nanoseconds.
-         */
-        private int maxPackets = -1;
-
-        /**
-         * The time stamp in nanoseconds of the start of the current
-         * <tt>perNanos</tt> interval.
-         */
-        private long millisStartTime = 0;
-
-        /**
-         * The list of RTP packets to be sent through the
-         * <tt>DatagramSocket</tt> of this <tt>OutputDataSource</tt>.
-         */
-        private final ArrayBlockingQueue<RawPacket> packetQueue
-            = new ArrayBlockingQueue<RawPacket>(
-                    MAX_PACKETS_PER_MILLIS_POLICY_PACKET_QUEUE_CAPACITY);
-
-        /**
-         * The number of RTP packets already sent during the current
-         * <tt>perNanos</tt> interval.
-         */
-        private long packetsSentInMillis = 0;
-
-        /**
-         * The time interval in nanoseconds during which {@link #maxPackets}
-         * number of RTP packets are to be sent through the
-         * <tt>DatagramSocket</tt> of this <tt>OutputDataSource</tt>.
-         */
-        private long perNanos = -1;
-
-        /**
-         * The <tt>Thread</tt> which is to send the RTP packets in
-         * {@link #packetQueue} through the <tt>DatagramSocket</tt> of this
-         * <tt>OutputDataSource</tt>.
-         */
-        private Thread sendThread;
-
-        /**
-         * The indicator which determines whether {@link #close()} has been
-         * invoked on this instance.
-         */
-        private boolean closed = false;
-
-        /**
-         * Initializes a new <tt>MaxPacketsPerMillisPolicy</tt> instance which
-         * is to control how many RTP packets this <tt>OutputDataSource</tt> is
-         * to send through its <tt>DatagramSocket</tt> per a specific number of
-         * milliseconds.
-         *
-         * @param maxPackets the maximum number of RTP packets to be sent per
-         * <tt>perMillis</tt> milliseconds through the <tt>DatagramSocket</tt>
-         * of this <tt>OutputDataStream</tt>
-         * @param perMillis the number of milliseconds per which a maximum of
-         * <tt>maxPackets</tt> RTP packets are to be sent through the
-         * <tt>DatagramSocket</tt> of this <tt>OutputDataStream</tt>
-         */
-        public MaxPacketsPerMillisPolicy(int maxPackets, long perMillis)
-        {
-            setMaxPacketsPerMillis(maxPackets, perMillis);
-            synchronized (this) {
-                if (sendThread == null)
-                {
-                    sendThread
-                        = new Thread(getClass().getName())
-                        {
-                            @Override
-                            public void run()
-                            {
-                                runInSendThread();
-                            }
-                        };
-                    sendThread.setDaemon(true);
-                    sendThread.start();
-                }
-            }
-        }
-
-        /**
-         * Closes the connector.
-         */
-        synchronized void close()
-        {
-            if (!closed)
-            {
-                closed = true;
-                // just offer a new packet to wakeup thread in case it waits for
-                // a packet.
-                packetQueue.offer(new RawPacket());
-            }
-        }
-
-        /**
-         * Sends the RTP packets in {@link #packetQueue} in accord with
-         * {@link #maxPackets} and {@link #perNanos}.
-         */
-        private void runInSendThread()
-        {
-            try
-            {
-                while (!closed)
-                {
-                    RawPacket packet;
-
-                    try
-                    {
-                        packet = packetQueue.take();
-                    }
-                    catch (InterruptedException iex)
-                    {
-                        continue;
-                    }
-                    if (closed)
-                        break;
-
-                    long time = System.nanoTime();
-                    long millisRemainingTime = time - millisStartTime;
-
-                    if ((perNanos < 1)
-                            || (millisRemainingTime >= perNanos))
-                    {
-                        millisStartTime = time;
-                        packetsSentInMillis = 0;
-                    }
-                    else if ((maxPackets > 0)
-                            && (packetsSentInMillis >= maxPackets))
-                    {
-                        while (true)
-                        {
-                            millisRemainingTime = System.nanoTime()
-                                    - millisStartTime;
-                            if (millisRemainingTime >= perNanos)
-                                break;
-                            LockSupport.parkNanos(millisRemainingTime);
-                        }
-                        millisStartTime = System.nanoTime();
-                        packetsSentInMillis = 0;
-                    }
-
-                    send(packet);
-                    packetsSentInMillis++;
-                }
-            }
-            finally
-            {
-                packetQueue.clear();
-                synchronized (packetQueue)
-                {
-                    if (Thread.currentThread().equals(sendThread))
-                        sendThread = null;
-                }
-            }
-        }
-
-        /**
-         * Sets the maximum number of RTP packets to be sent by this
-         * <tt>OutputDataStream</tt> through its <tt>DatagramSocket</tt> per
-         * a specific number of milliseconds.
-         *
-         * @param maxPackets the maximum number of RTP packets to be sent by
-         * this <tt>OutputDataStream</tt> through its <tt>DatagramSocket</tt>
-         * per the specified number of milliseconds; <tt>-1</tt> if no maximum
-         * is to be set
-         * @param perMillis the number of milliseconds per which
-         * <tt>maxPackets</tt> are to be sent by this <tt>OutputDataStream</tt>
-         * through its <tt>DatagramSocket</tt>
-         */
-        public void setMaxPacketsPerMillis(int maxPackets, long perMillis)
-        {
-            if (maxPackets < 1)
-            {
-                this.maxPackets = -1;
-                this.perNanos = -1;
-            }
-            else
-            {
-                if (perMillis < 1)
-                    throw new IllegalArgumentException("perMillis");
-
-                this.maxPackets = maxPackets;
-                this.perNanos = perMillis * 1000000;
-            }
-        }
-
-        /**
-         * Queues a specific RTP packet to be sent through the
-         * <tt>DatagramSocket</tt> of this <tt>OutputDataStream</tt>.
-         *
-         * @param packet the RTP packet to be queued for sending through the
-         * <tt>DatagramSocket</tt> of this <tt>OutputDataStream</tt>
-         */
-        public void write(RawPacket packet)
-        {
-            do
-            {
-                try
-                {
-                    packetQueue.put(packet);
-                    break;
-                }
-                catch (InterruptedException iex)
-                {
-                }
-            }
-            while (true);
-        }
-    }
-
-    /**
-     * Enables or disables this <tt>RTPConnectorOutputStream</tt>.
-     * While the stream is disabled, it suppresses actually sending any packets
-     * via {@link #send(RawPacket)}.
-     *
-     * @param enabled <tt>true</tt> to enable, <tt>false</tt> to disable.
-     */
-    public void setEnabled(boolean enabled)
-    {
-        if (logger.isDebugEnabled())
-            logger.debug("setEnabled: " + enabled);
-
-        this.enabled = enabled;
     }
 }
